refactor: serialization functions
